@@ -4,17 +4,13 @@
   "description": "A minimal node-pty Electron example",
   "main": "main.js",
   "scripts": {
-    "start": "electron ."
+    "start": "electron . --ignore-gpu-blacklist"
   },
   "repository": "https://github.com/Tyriar/node-pty",
   "author": "Tyriar",
-  "devDependencies": {
-    "electron": "1.8.4"
-  },
   "dependencies": {
-    "node-addon-api": "1.7.1",
-    "node-addon-api": "1.7.1",
-    "node-pty": "^0.7.4",
-    "xterm": "^3.0.2"
+    "electron": "^9.1.0",
+    "node-pty": "^0.9.0",
+    "xterm": "4.2.0"
   }
 }
