@@ -2,7 +2,7 @@
   "name": "profoundjs-node-pty",
   "description": "Fork pseudoterminals in Node.JS",
   "author": {
-    "name": "Daniel Imms"
+    "name": "Microsoft Corporation"
   },
   "version": "1.0.5",
   "license": "MIT",
@@ -18,7 +18,7 @@
     "scripts/",
     "src/",
     "deps/",
-    "bin/"
+    "typings/"
   ],
   "homepage": "https://github.com/Tyriar/node-pty",
   "bugs": {
@@ -33,23 +33,28 @@
     "openpty"
   ],
   "scripts": {
-    "tsc": "tsc",
-    "lint": "tslint src/**/*.ts",
+    "build": "tsc -b ./src/tsconfig.json",
+    "watch": "tsc -b -w ./src/tsconfig.json",
+    "lint": "eslint -c .eslintrc.js --ext .ts src/",
     "install": "node scripts/install.js",
     "postinstall": "node scripts/post-install.js",
     "test": "cross-env NODE_ENV=test mocha -R spec --exit lib/*.test.js",
-    "prepare": "npm run tsc"
+    "posttest": "npm run lint",
+    "prepare": "npm run build",
+    "prepublishOnly": "npm run build"
   },
   "dependencies": {
-    "node-addon-api": "1.7.1"
+    "node-addon-api": "^3.0.2"
   },
   "devDependencies": {
-    "@types/mocha": "^5.0.0",
-    "@types/node": "^6.0.104",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "8",
+    "@typescript-eslint/eslint-plugin": "^2.27.0",
+    "@typescript-eslint/parser": "^2.27.0",
     "cross-env": "^5.1.4",
-    "mocha": "^5.0.5",
-    "pollUntil": "^1.0.3",
-    "tslint": "^5.9.1",
-    "typescript": "^2.8.1"
+    "eslint": "^6.8.0",
+    "mocha": "^7.1.1",
+    "ps-list": "^6.0.0",
+    "typescript": "^3.8.3"
   }
 }
