@@ -1,11 +1,13 @@
 /**
  * Copyright (c) 2017, Daniel Imms (MIT License).
+ * Copyright (c) 2018, Microsoft Corporation (MIT License).
  */
 
 import { UnixTerminal } from './unixTerminal';
 import * as assert from 'assert';
-import pollUntil = require('pollUntil');
+import * as cp from 'child_process';
 import * as path from 'path';
+import { pollUntil } from './testUtils.test';
 
 const FIXTURES_PATH = path.normalize(path.join(__dirname, '..', 'fixtures', 'utf8-character.txt'));
 
@@ -14,7 +16,7 @@ if (process.platform !== 'win32') {
     describe('Constructor', () => {
       it('should set a valid pts name', () => {
         const term = new UnixTerminal('/bin/bash', [], {});
-        let regExp;
+        let regExp: RegExp;
         if (process.platform === 'linux') {
           // https://linux.die.net/man/4/pts
           regExp = /^\/dev\/pts\/\d+$/;
@@ -26,7 +28,7 @@ if (process.platform !== 'win32') {
         if (regExp) {
           assert.ok(regExp.test((<any>term)._pty), '"' + (<any>term)._pty + '" should match ' + regExp.toString());
         }
-        term.kill();  //Make sure one forked pty isn't still doing waitpid.
+        term.kill();
       });
     });
 
@@ -41,7 +43,7 @@ if (process.platform !== 'win32') {
       });
       it('should return a Buffer when encoding is null', (done) => {
         const term = new UnixTerminal('/bin/bash', [ '-c', `cat "${FIXTURES_PATH}"` ], {
-          encoding: null,
+          encoding: null
         });
         term.on('data', (data) => {
           assert.equal(typeof data, 'object');
@@ -62,7 +64,7 @@ if (process.platform !== 'win32') {
           buffer += data;
         });
         term.on('exit', () => {
-          assert.equal(new Buffer(buffer, 'base64').toString().replace('\r', '').replace('\n', ''), text);
+          assert.equal(Buffer.alloc(8, buffer, 'base64').toString().replace('\r', '').replace('\n', ''), text);
           done();
         });
       });
@@ -79,7 +81,6 @@ if (process.platform !== 'win32') {
       });
 
       it('should open a pty with access to a master and slave socket', (done) => {
-        let doneCalled = false;
         term = UnixTerminal.open({});
 
         let slavebuf = '';
@@ -92,27 +93,152 @@ if (process.platform !== 'win32') {
           masterbuf += data;
         });
 
-        (<any>pollUntil)(() => {
+        pollUntil(() => {
           if (masterbuf === 'slave\r\nmaster\r\n' && slavebuf === 'master\n') {
             done();
             return true;
           }
           return false;
-        }, [], 200, 10);
+        }, 200, 10);
 
         term.slave.write('slave\n');
         term.master.write('master\n');
       });
-
-      it('onexit fires', (done) => {
-        const term = new UnixTerminal('/bin/bash', [ '-c', `cat "${FIXTURES_PATH}"` ], {
-          encoding: null,
+    });
+    describe('signals in parent and child', () => {
+      it('SIGINT - custom in parent and child', done => {
+        // this test is cumbersome - we have to run it in a sub process to
+        // see behavior of SIGINT handlers
+        const data = `
+        var pty = require('./lib/index');
+        process.on('SIGINT', () => console.log('SIGINT in parent'));
+        var ptyProcess = pty.spawn('node', ['-e', 'process.on("SIGINT", ()=>console.log("SIGINT in child"));setTimeout(() => null, 300);'], {
+          name: 'xterm-color',
+          cols: 80,
+          rows: 30,
+          cwd: process.env.HOME,
+          env: process.env
         });
-        term.on('exit', () => {
+        ptyProcess.on('data', function (data) {
+          console.log(data);
+        });
+        setTimeout(() => null, 500);
+        console.log('ready', ptyProcess.pid);
+        `;
+        const buffer: string[] = [];
+        const p = cp.spawn('node', ['-e', data]);
+        let sub = '';
+        p.stdout.on('data', (data) => {
+          if (!data.toString().indexOf('ready')) {
+            sub = data.toString().split(' ')[1].slice(0, -1);
+            setTimeout(() => {
+              process.kill(parseInt(sub), 'SIGINT');  // SIGINT to child
+              p.kill('SIGINT');                       // SIGINT to parent
+            }, 200);
+          } else {
+            buffer.push(data.toString().replace(/^\s+|\s+$/g, ''));
+          }
+        });
+        p.on('close', () => {
+          // handlers in parent and child should have been triggered
+          assert.equal(buffer.indexOf('SIGINT in child') !== -1, true);
+          assert.equal(buffer.indexOf('SIGINT in parent') !== -1, true);
           done();
         });
       });
+      it('SIGINT - custom in parent, default in child', done => {
+        // this tests the original idea of the signal(...) change in pty.cc:
+        // to make sure the SIGINT handler of a pty child is reset to default
+        // and does not interfere with the handler in the parent
+        const data = `
+        var pty = require('./lib/index');
+        process.on('SIGINT', () => console.log('SIGINT in parent'));
+        var ptyProcess = pty.spawn('node', ['-e', 'setTimeout(() => console.log("should not be printed"), 300);'], {
+          name: 'xterm-color',
+          cols: 80,
+          rows: 30,
+          cwd: process.env.HOME,
+          env: process.env
+        });
+        ptyProcess.on('data', function (data) {
+          console.log(data);
+        });
+        setTimeout(() => null, 500);
+        console.log('ready', ptyProcess.pid);
+        `;
+        const buffer: string[] = [];
+        const p = cp.spawn('node', ['-e', data]);
+        let sub = '';
+        p.stdout.on('data', (data) => {
+          if (!data.toString().indexOf('ready')) {
+            sub = data.toString().split(' ')[1].slice(0, -1);
+            setTimeout(() => {
+              process.kill(parseInt(sub), 'SIGINT');  // SIGINT to child
+              p.kill('SIGINT');                       // SIGINT to parent
+            }, 200);
+          } else {
+            buffer.push(data.toString().replace(/^\s+|\s+$/g, ''));
+          }
+        });
+        p.on('close', () => {
+          // handlers in parent and child should have been triggered
+          assert.equal(buffer.indexOf('should not be printed') !== -1, false);
+          assert.equal(buffer.indexOf('SIGINT in parent') !== -1, true);
+          done();
+        });
+      });
+      it('SIGHUP default (child only)', done => {
+        const term = new UnixTerminal('node', [ '-e', `
+        console.log('ready');
+        setTimeout(()=>console.log('timeout'), 200);`
+        ]);
+        let buffer = '';
+        term.on('data', (data) => {
+          if (data === 'ready\r\n') {
+            term.kill();
+          } else {
+            buffer += data;
+          }
+        });
+        term.on('exit', () => {
+          // no timeout in buffer
+          assert.equal(buffer, '');
+          done();
+        });
+      });
+      it('SIGUSR1 - custom in parent and child', done => {
+        let pHandlerCalled = 0;
+        const handleSigUsr = function(h: any): any {
+          return function(): void {
+            pHandlerCalled += 1;
+            process.removeListener('SIGUSR1', h);
+          };
+        };
+        process.on('SIGUSR1', handleSigUsr(handleSigUsr));
 
+        const term = new UnixTerminal('node', [ '-e', `
+        process.on('SIGUSR1', () => {
+          console.log('SIGUSR1 in child');
+        });
+        console.log('ready');
+        setTimeout(()=>null, 200);`
+        ]);
+        let buffer = '';
+        term.on('data', (data) => {
+          if (data === 'ready\r\n') {
+            process.kill(process.pid, 'SIGUSR1');
+            term.kill('SIGUSR1');
+          } else {
+            buffer += data;
+          }
+        });
+        term.on('exit', () => {
+          // should have called both handlers and only once
+          assert.equal(pHandlerCalled, 1);
+          assert.equal(buffer, 'SIGUSR1 in child\r\n');
+          done();
+        });
+      });
     });
   });
 }
