@@ -6,68 +6,57 @@ const path = require('path');
 const spawn = require('child_process').spawn;
 
 // IBM i uses a different module.
-if (os.type() == "OS400")
+if (os.type() == 'OS400')
   process.exit(0);
 
 // Test for pre-built binary.
-const NODE_MODULE_VERSION = parseInt(process.versions.modules, 10);
-var binPath = `${process.platform}-${process.arch}`;  //N-API Binaries for Node versions 10 and later are in folders like win32-x64/
-if (NODE_MODULE_VERSION < 64 || ['darwin-x64','linux-ppc64','linux-x64','win32-x64'].indexOf(binPath) < 0 ){
-  binPath += `-${process.versions.modules}`;    //NAN Binaries for Node versions before 10 are in folders like win32-x64-57/.
-}
-
+const binPath = `${process.platform}-${process.arch}`;
+const binNames = process.platform === "win32" ? ["conpty.node", "conpty_console_list.node", "pty.node", "winpty.dll", "winpty-agent.exe"] : ["pty.node"];
 try {
-  
-  fs.statSync(path.join(__dirname, '..', 'bin', binPath, 'pty.node'));
+
+  binNames.forEach(binName => {
+    fs.statSync(path.join(__dirname, '..', 'bin', binPath, binName));
+  });
   console.log(`${binPath} exists, exiting`);
-  
+
 }
 catch (error) {
-  
+
   build();
-  
+
 }
 
 function build() {
-  
-  const p = spawn(os.platform() === 'win32' ? 'node-gyp.cmd' : 'node-gyp', ['rebuild'], {
+
+  const gypArgs = ['rebuild'];
+  if (process.env.NODE_PTY_DEBUG) {
+    gypArgs.push('--debug');
+  }
+  const gypProcess = spawn(os.platform() === 'win32' ? 'node-gyp.cmd' : 'node-gyp', gypArgs, {
     cwd: path.join(__dirname, '..'),
     stdio: 'inherit'
   });
-
-  p.on('exit', function (code) {
-    
-    if (code == 0)
+  gypProcess.on('exit', function (code) {
+    if (code === 0)
       afterBuild();
     process.exit(code);
-    
   });
 
 }
 
 function afterBuild() {
-  
+
   try {
-    
+
     fs.mkdirSync(path.join(__dirname, '..', 'bin', binPath));
-    
+
   }
   catch (error) {}
-  fs.copyFileSync(
-    path.join(__dirname, '..', 'build', 'Release', 'pty.node'),
-    path.join(__dirname, '..', 'bin', binPath, 'pty.node')
-  );
-  if (os.platform() == 'win32') {
-    
+  binNames.forEach(binName => {
     fs.copyFileSync(
-      path.join(__dirname, '..', 'build', 'Release', 'winpty.dll'),
-      path.join(__dirname, '..', 'bin', binPath, 'winpty.dll')
+      path.join(__dirname, '..', 'build', 'Release', binName),
+      path.join(__dirname, '..', 'bin', binPath, binName)
     );
-    fs.copyFileSync(
-      path.join(__dirname, '..', 'build', 'Release', 'winpty-agent.exe'),
-      path.join(__dirname, '..', 'bin', binPath, 'winpty-agent.exe')
-    );
-  
-  }
-  
+  });
+
 }
